qna 발표 전 최종버전

+++ src/main/java/shop/heartmuscle/heartmuscle/service/QnaCommentService.java
@@ -1,10 +1,10 @@
 package shop.heartmuscle.heartmuscle.service;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import shop.heartmuscle.heartmuscle.domain.Qna;
 import shop.heartmuscle.heartmuscle.domain.QnaComment;
+import shop.heartmuscle.heartmuscle.domain.User;
 import shop.heartmuscle.heartmuscle.dto.QnaCommentRequestDto;
+import shop.heartmuscle.heartmuscle.dto.ResultResponseDto;
 import shop.heartmuscle.heartmuscle.repository.QnaCommentRepository;
 import shop.heartmuscle.heartmuscle.repository.QnaRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,8 +12,6 @@ import org.springframework.stereotype.Service;
 import shop.heartmuscle.heartmuscle.security.UserDetailsImpl;
 
 import javax.transaction.Transactional;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +21,7 @@ public class QnaCommentService {
     private final QnaRepository qnaRepository;
 
     @Transactional
-    public void createQnaComment(Long qnaId, QnaCommentRequestDto qnaCommentRequestDto, UserDetailsImpl nowUser){
+    public void createQnaComment(Long qnaId, QnaCommentRequestDto qnaCommentRequestDto, UserDetailsImpl nowUser) {
         Qna qna = qnaRepository.findById(qnaId).orElseThrow(
                 () -> new NullPointerException("댓글못단다고오오오")
         );
@@ -31,30 +29,54 @@ public class QnaCommentService {
         qnaCommentRepository.save(qnaComment);
     }
 
-
+    //댓글 수정
     @Transactional
-    public QnaComment updateQnaComment(Long qnaId, Long qnaCommentId, QnaCommentRequestDto qnaCommentDto, UserDetailsImpl nowUser) {
-        qnaRepository.findById(qnaId).orElseThrow(
-                () -> new NullPointerException("해당 리뷰가 존재하지 않습니다."));
+    public ResultResponseDto updateQnaComment(Long qnaId, Long qnaCommentId, QnaCommentRequestDto qnaCommentDto, UserDetailsImpl nowUser) {
 
-        QnaComment qnaComment = qnaCommentRepository.findById(qnaCommentId).orElseThrow(
-                () -> new NullPointerException("해당 댓글이 존재하지 않습니다."));
+        Long nowUserId = nowUser.getId();
 
-        qnaComment.update(qnaCommentDto.getComment());
-        return qnaComment;
-    }
-
-    // 댓글 삭제
-    @Transactional
-    public ResponseEntity<?> deleteQnaComment(Long qnaId, Long qnaCommentId, UserDetailsImpl nowUser) {
         qnaRepository.findById(qnaId).orElseThrow(
                 () -> new NullPointerException("해당 게시글이 존재하지 않습니다.")
         );
         QnaComment qnaComment = qnaCommentRepository.findById(qnaCommentId).orElseThrow(
                 () -> new NullPointerException("해당 댓글이 존재하지 않습니다.")
         );
-        qnaCommentRepository.delete(qnaComment);
-        return new ResponseEntity<>(HttpStatus.OK);
+
+        User findQnaCommentUser = qnaComment.getUser();
+        Long QnaCommentUser = findQnaCommentUser.getId();
+
+        if (nowUserId == QnaCommentUser) {
+            qnaComment.update(qnaCommentDto.getComment());
+            return new ResultResponseDto("success", "댓글이 수정 되었습니다.");
+        } else {
+            return new ResultResponseDto("fail", "댓글 수정 권한이 없습니다. .");
+        }
     }
 
+    // 댓글 삭제
+    @Transactional
+    public ResultResponseDto deleteQnaComment(Long qnaId, Long qnaCommentId, UserDetailsImpl nowUser) {
+
+        Long nowUserId = nowUser.getId();
+
+        qnaRepository.findById(qnaId).orElseThrow(
+                () -> new NullPointerException("해당 게시글이 존재하지 않습니다.")
+        );
+        QnaComment qnaComment = qnaCommentRepository.findById(qnaCommentId).orElseThrow(
+                () -> new NullPointerException("해당 댓글이 존재하지 않습니다.")
+        );
+
+        User findQnaCommentUser = qnaComment.getUser();
+        Long QnaCommentUser = findQnaCommentUser.getId();
+
+        if (nowUserId == QnaCommentUser) {
+            qnaCommentRepository.delete(qnaComment);
+            return new ResultResponseDto("success", "댓글이 삭제 되었습니다.");
+        } else {
+            return new ResultResponseDto("fail", "댓글 삭제 권한이 없습니다. .");
+        }
+
+    }
 }
+
+
